Expose "cullude" to cgame and completely skip occluded misc_dust.
