+++ src/cgame/cgame.h
@@ -738,6 +738,21 @@ typedef struct cg_import_s {
 	void (*InitView)(r_view_t *view);
 
 	/**
+	 * @brief Add an occlusion query to the scene.
+	 */
+	void (*AddOcclusionQuery)(r_view_t *view, const box3_t bounds);
+
+	/**
+	 * @return True if the bounding box is culled or occluded, false otherwise.
+	 */
+	_Bool (*CulludeBox)(const r_view_t *view, const box3_t bounds);
+
+	/**
+	 * @brief True if the bounding sphere is culled or occluded, false otherwise.
+	 */
+	_Bool (*CulludeSphere)(const r_view_t *view, const vec3_t origin, float radius);
+
+	/**
 	 * @brief Adds an entity to the scene for the current frame.
 	 * @return The added entity.
 	 */
@@ -762,11 +777,6 @@ typedef struct cg_import_s {
 	 * @brief Add a stain to the scene.
 	 */
 	void (*AddStain)(r_view_t *view, const r_stain_t *s);
-
-	/**
-	 * @brief Add an occlusion query to the scene.
-	 */
-	void (*AddOcclusionQuery)(r_view_t *view, const box3_t bounds);
 
 	/**
 	 * @brief Draws the player model view.
